remove unnecessary mock tests
